requirements.txt
torch>=2.4
torchvision>=0.19
diffusers>=0.30
transformers>=4.45
accelerate>=0.34
gradio>=5.0
huggingface_hub>=0.30
Pillow>=10.0
numpy>=1.26
gdown>=5.1
kaggle>=1.6
PyYAML>=6.0
audioop-lts>=0.2; python_version >= "3.13"
